chore: move constants to a constants folder
