+++ src/cli/core/constants/regularClassGroupKeys.ts
@@ -1,48 +1,3 @@
-export const allVariants = [
-  'responsive',
-  'motion-safe',
-  'motion-reduce',
-  'first',
-  'last',
-  'odd',
-  'even',
-  'visited',
-  'checked',
-  'group-hover',
-  'group-focus',
-  'focus-within',
-  'hover',
-  'focus',
-  'focus-visible',
-  'active',
-  'disabled',
-  // Exhaustive pseudo-classess
-  'only',
-  'first-of-type',
-  'last-of-type',
-  'only-of-type',
-  'target',
-  'default',
-  'indeterminate',
-  'placeholder-shown',
-  'autofill',
-  'required',
-  'valid',
-  'invalid',
-  'in-range',
-  'out-of-range',
-  // New peer-*, selection & marker variants and before/after
-  'peer-hover',
-  'peer-checked',
-  'peer-focus',
-  'selection',
-  'marker',
-  'before',
-  'after',
-  // dark mode utility
-  'dark',
-];
-
 export const regularClassGroupKeys = [
   // 'accessibility',
   'screenReaders',
